Volume Based Density Added FindNearbyPointsInDimension for Octree Cell's. Updated IEntitySearchTree with Area search by Vector3 dimension. Added bounding area search for entites. Updated GetMapNodesAroundPositionEvent to constructor based. Moved AStar Search logic to encapsulated class. Formatting and cleanup.

+++ src/EventHorizon.Game.Server.Zone.Model/Map/MapEdge.cs
@@ -2,12 +2,11 @@ namespace EventHorizon.Game.Server.Zone.Model.Map
 {
     public struct MapEdge
     {
-        // TODO: Add test to make sure Path finding still works with default(MapEdge);
         public static MapEdge NULL = new MapEdge(-1, -1);
 
         public int FromIndex { get; set; }
         public int ToIndex { get; set; }
-        public int Cost { get; set; }
+        public float Cost { get; set; }
 
         public MapEdge(int fromIndex, int toIndex)
         {
